chore(ipld): move out validation to separate files and fixes from review by @evan-forbes
