+++ p2p/ipld/sample.go
@@ -16,14 +16,14 @@ type Sample struct {
 
 // SampleSquare randomly picks *num* unique points from arbitrary *width* square
 // and returns them as samples.
-func SampleSquare(squareWidth uint32, num int) []*Sample {
+func SampleSquare(squareWidth uint32, num int) []Sample {
 	ss := newSquareSampler(squareWidth, num)
 	ss.sample(num)
 	return ss.sampled()
 }
 
 // Leaf returns leaf info needed for retrieval using data provided with DAHeader.
-func (s *Sample) Leaf(dah *types.DataAvailabilityHeader) (cid.Cid, uint32, error) {
+func (s Sample) Leaf(dah *types.DataAvailabilityHeader) (cid.Cid, uint32, error) {
 	var (
 		leaf uint32
 		root namespace.IntervalDigest
@@ -47,53 +47,52 @@ func (s *Sample) Leaf(dah *types.DataAvailabilityHeader) (cid.Cid, uint32, error
 }
 
 // Equals check whenever to samples are equal.
-func (s *Sample) Equals(to *Sample) bool {
+func (s Sample) Equals(to Sample) bool {
 	return s.Row == to.Row && s.Col == to.Col
 }
 
 type squareSampler struct {
 	squareWidth uint32
-	samples     []*Sample
+	samples     map[Sample]struct{}
 }
 
 func newSquareSampler(squareWidth uint32, expectedSamples int) *squareSampler {
 	return &squareSampler{
 		squareWidth: squareWidth,
-		samples:     make([]*Sample, 0, expectedSamples),
+		samples:     make(map[Sample]struct{}, expectedSamples),
 	}
 }
 
 func (ss *squareSampler) sample(num int) {
 	done := 0
 	for done < num {
-		p := &Sample{
+		s := Sample{
 			Row: uint32(rand.Int31n(int32(ss.squareWidth))),
 			Col: uint32(rand.Int31n(int32(ss.squareWidth))),
 		}
 
-		if ss.isSampled(p) {
+		if ss.isSampled(s) {
 			continue
 		}
 
 		done++
-		ss.addSample(p)
+		ss.addSample(s)
 	}
 }
 
-func (ss *squareSampler) sampled() []*Sample {
-	return ss.samples
-}
-
-func (ss *squareSampler) addSample(p *Sample) {
-	ss.samples = append(ss.samples, p)
-}
-
-func (ss *squareSampler) isSampled(p *Sample) bool {
-	for _, sp := range ss.samples {
-		if sp.Equals(p) {
-			return true
-		}
+func (ss *squareSampler) sampled() []Sample {
+	samples := make([]Sample, 0, len(ss.samples))
+	for s := range ss.samples {
+		samples = append(samples, s)
 	}
+	return samples
+}
 
-	return false
+func (ss *squareSampler) addSample(s Sample) {
+	ss.samples[s] = struct{}{}
+}
+
+func (ss *squareSampler) isSampled(s Sample) bool {
+	_, ok := ss.samples[s]
+	return ok
 }
